Fixes #36 - Repeated lines are detected and erased.
Ludens (from FICS) told us that repeated lines are expected. Then we filter
repeated lines by checking turn color. If previos color is the same as current
then we consider it a repeated line.

Code refactorization was done. Communication protocol with LPT was moved to
comms.py and code was cleaned.